# Stellify-Update/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stellify Church Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with PCM/TCM ID and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Verify identity (churchId + email) and issue a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Identity verified, reset token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetTokenResponse"
        "400":
          description: Missing fields
        "404":
          description: No matching account
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "204":
          description: Password reset
        "400":
          description: Invalid or expired token

  # ─── MEMBERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [churchId, password]
      properties:
        churchId:
          type: string
          description: PCM0001 or TCM0001 format
        password:
          type: string
    AuthResult:
      type: object
      required: [member, token]
      properties:
        member:
          $ref: "#/components/schemas/Member"
        token:
          type: string
        passwordChangeRequired:
          type: boolean
          description: When true, member must change their password before accessing the app
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        memberId:
          type: ["integer", "null"]
          description: Admin can reset for another member
    Member:
      type: object
      required: [id, churchId, accountType, status, role, firstName, lastName, discipleshipEnabled, stellarStatus, isTestingAccount]
      properties:
        id:
          type: integer
        churchId:
          type: string
        accountType:
          type: string
          enum: [PCM, TCM]
        status:
          type: string
          enum: [Active, Deactivated, Retired, Deleted]
        role:
          type: string
          enum: [Admin, Pastor, Servant, SharingCaptain, CgAuditor, ChurchAuditor, Member]
        firstName:
          type: string
        middleName:
          type: ["string", "null"]
        lastName:
          type: string
        nickname:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [Male, Female, null]
        birthday:
          type: ["string", "null"]
        civilStatus:
          type: ["string", "null"]
        weddingAnniversary:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        discipleshipEnabled:
          type: boolean
        sbgEndDate:
          type: ["string", "null"]
        stellarStatus:
          type: string
          enum: [SBG, P2S, S2B, P2G]
        pendingStellarStatus:
          type: ["string", "null"]
          description: Status awaiting pastor approval
        stellarStatusApprovedByPastor:
          type: boolean
        isTestingAccount:
          type: boolean
          description: Testing accounts are excluded from the real StellarBoard
        mustChangePassword:
          type: boolean
          description: When true, member is required to change their password on next login
        familyGroupNumber:
          type: ["string", "null"]
        familyRole:
          type: ["string", "null"]
        cgNumber:
          type: ["string", "null"]
        cgRole:
          type: ["string", "null"]
        sharingHuddleNumber:
          type: ["string", "null"]
        sharingHuddleRole:
          type: ["string", "null"]
        salvationDate:
          type: ["string", "null"]
        baptismDate:
          type: ["string", "null"]
        sharingPermissionDate:
          type: ["string", "null"]
        bearingDate:
          type: ["string", "null"]
        dedicationDate:
          type: ["string", "null"]
        ministryRoles:
          type: ["array", "null"]
          items:
            type: string
          description: Array of additional leadership roles held by this member
        roleGroupMappings:
          type: ["object", "null"]
          additionalProperties:
            type: string
          description: Maps each leadership role to its assigned group number
        foodPreferences:
          type: ["string", "null"]
        medicalCondition:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactRelationship:
          type: ["string", "null"]
        emergencyContactNumber:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
          description: IANA timezone string (e.g. "Asia/Manila"). Stored by the client on login and used for server-side streak-boundary and reminder calculations.
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ForgotPasswordInput:
      type: object
      required: [churchId, email]
      properties:
        churchId:
          type: string
        email:
          type: string
    PasswordResetTokenResponse:
      type: object
      required: [resetToken]
      properties:
        resetToken:
          type: string
    ResetPasswordInput:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
        newPassword:
          type: string
          minLength: 6
...